chore: move admin schemas in dedicated admin schema folder
